Applied fixes suggested by Yovanha
